refactor: 회원가입관련 컴포넌트 역할 명확화 및 로직 개선
