+++ src/app/(providers)/(auth)/sign-up/_components/EssentialInfoForm/EssentialInfoForm.tsx
@@ -1,10 +1,121 @@
-import { EssentialInfoFormProps } from '@/types/auth';
+'use client';
 
-const EssentialInfoForm = ({ formState, handleChange, handleCheckDuplicate, onSubmit }: EssentialInfoFormProps) => {
+import { EssentialInfoFormProps, FormState } from '@/types/auth';
+import { useState } from 'react';
+import { validatePassword } from '../../../_utils/passwordValidation';
+
+const EssentialInfoForm = ({ formState, setFormState, checkDuplicate, onNext }: EssentialInfoFormProps) => {
+  const [isCheckingDuplicate, setIsCheckingDuplicate] = useState({
+    email: false,
+    nickname: false,
+  });
+
+  const validationRules = {
+    email: (value: string) => (!value.includes('@') ? '유효한 이메일 주소를 입력해주세요.' : null),
+    nickname: (value: string) => (value.length < 2 ? '닉네임은 2자 이상이어야 합니다.' : null),
+    password: validatePassword,
+    confirmPassword: (value: string, password: string) => (value !== password ? '비밀번호가 일치하지 않습니다.' : null),
+  };
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target;
+    setFormState((prev) => ({
+      ...prev,
+      [name]: {
+        ...prev[name as keyof FormState],
+        value,
+        error: null,
+        successMessage: null,
+        isVerified: false,
+      },
+    }));
+  };
+
+  const handleCheckDuplicate = async (field: 'email' | 'nickname') => {
+    if (!formState[field].value.trim()) {
+      setFormState((prev) => ({
+        ...prev,
+        [field]: {
+          ...prev[field],
+          error: `${field === 'email' ? '이메일' : '닉네임'}을 입력해주세요.`,
+          successMessage: null,
+          isVerified: false,
+        },
+      }));
+      return;
+    }
+
+    setIsCheckingDuplicate((prev) => ({ ...prev, [field]: true }));
+
+    try {
+      const result = await checkDuplicate(field, formState[field].value);
+
+      setFormState((prev) => ({
+        ...prev,
+        [field]: {
+          ...prev[field],
+          error: result ? null : `이미 사용 중인 ${field === 'email' ? '이메일' : '닉네임'}입니다.`,
+          successMessage: result ? `사용 가능한 ${field === 'email' ? '이메일' : '닉네임'}입니다.` : null,
+          isVerified: result,
+        },
+      }));
+    } catch (error) {
+      console.error(`${field} 중복 확인 중 오류 발생:`, error);
+      setFormState((prev) => ({
+        ...prev,
+        [field]: {
+          ...prev[field],
+          error: '중복 확인 중 오류가 발생했습니다.',
+        },
+      }));
+    } finally {
+      setIsCheckingDuplicate((prev) => ({ ...prev, [field]: false }));
+    }
+  };
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+
+    let hasError = false;
+    const newFormState = { ...formState };
+
+    // 모든 필수 필드 유효성 검사
+    for (const field of ['email', 'nickname', 'password', 'confirmPassword'] as const) {
+      let error = null;
+      if (field === 'confirmPassword') {
+        error = validationRules[field](newFormState[field].value, newFormState.password.value);
+      } else {
+        error = validationRules[field](newFormState[field].value);
+      }
+
+      if (error) {
+        newFormState[field].error = error;
+        hasError = true;
+      }
+    }
+
+    // 중복 확인 여부 체크
+    if (!newFormState.email.isVerified) {
+      newFormState.email.error = '이메일 중복 확인이 필요합니다.';
+      hasError = true;
+    }
+    if (!newFormState.nickname.isVerified) {
+      newFormState.nickname.error = '닉네임 중복 확인이 필요합니다.';
+      hasError = true;
+    }
+
+    if (hasError) {
+      setFormState(newFormState);
+      return;
+    }
+
+    onNext();
+  };
+
   return (
     <form
       className="flex flex-col gap-4 items-center justify-center w-full max-w-[390px] sm:max-w-[540px] md:max-w-[720px] lg:max-w-[960px] xl:max-w-[1140px]"
-      onSubmit={onSubmit}
+      onSubmit={handleSubmit}
     >
       <div className="flex flex-col items-center w-2/3 sm:w-3/5 md:w-2/5 lg:w-1/3 xl:w-1/4">
         <div className="w-full">
@@ -29,7 +140,7 @@ const EssentialInfoForm = ({ formState, handleChange, handleCheckDuplicate, onSu
               className="bg-[#D9D9D9] w-20 h-full text-nowrap rounded-md px-2.5 py-2 ml-2 hover:brightness-90"
               aria-label="이메일 중복 확인"
             >
-              {formState.email.isChecking ? '확인 중' : '확인'}
+              {isCheckingDuplicate.email ? '확인 중' : '확인'}
             </button>
           </div>
         </div>
@@ -62,7 +173,7 @@ const EssentialInfoForm = ({ formState, handleChange, handleCheckDuplicate, onSu
               className="bg-[#D9D9D9] w-20 h-full text-nowrap rounded-md px-2.5 py-2 ml-2 hover:brightness-90"
               aria-label="닉네임 중복 확인"
             >
-              {formState.nickname.isChecking ? '확인 중' : '확인'}
+              {isCheckingDuplicate.nickname ? '확인 중' : '확인'}
             </button>
           </div>
         </div>
